+++ tests/cluecode/data/authors/author_dot_joined_name-TM1637_cpp.cpp
@@ -1,0 +1,2 @@
+//  Date:9 April,2012
+//  Author:Frankie.Chu